+++ gagu-front-end/src/utils/app.util.ts
@@ -1,0 +1,10 @@
+import { IApp, IRunningApp } from '../types'
+
+export const genRunningApp = (app: IApp) => {
+  const _app: IRunningApp = {
+    ...app,
+    runningId: Date.now(),
+    visible: true,
+  }
+  return _app
+}